Backend/requirements.txt
flask
flask-cors
qrcode
pillow
gunicorn